ar2en-service/requirements.txt
flask==3.0.0
flask-restful==0.3.10
transformers==4.36.2
torch==2.5.1
python-dotenv==1.0.0
sentencepiece==0.1.99
gunicorn==21.2.0
